refactor: Change name to files in user module
