+++ src/user/types/inject.ts
@@ -1,4 +1,4 @@
-import { UserQueryRepository } from "../services/UserQuery.repository";
+import { UserQueryRepository } from "../services/user.repository";
 
 export type Service = {
     userQuery: UserQueryRepository;
