+++ src/user/views/user-detail/UserDetailView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { toRef } from 'vue';
-import { useUserDetailController } from './UserDetail.controller';
+import { useUserDetailController } from './userDetail.controller';
 import UserDetailSkeleton from './components/UserDetailSkeleton.vue';
 import UserDetailContainer from './components/UserDetailContainer.vue';
 
@@ -13,4 +13,4 @@ const { isLoading, user } = useUserDetailController(toRef(props, 'id'))
 <template>
     <UserDetailSkeleton v-if="isLoading" />
     <UserDetailContainer v-else :user="user" />
-</template>
+</template>./userDetail.controller
